requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
supabase==2.4.6
inngest==0.4.0
python-dotenv==1.0.1
pydantic==2.7.1
PyJWT[crypto]==2.8.0
